Restructured runtime logic and replaced bespoke type detection logic with reusable variant
